Authored by Kyle Norman and Javier Marin
Started writing insert function and helper functions. Including but not limited too:

Insert: Inserts an expression with a certain opetation to an already existing expression.

CheckIfInsertShouldGoAboveOrBelow: Checks whether the inserted expression can be safely balanced below the tree or
will become the new root of the tree.

Current Problems:

We need to fix the templating issue with Insert as there is no way to make a new BinaryExpression inisde of it right now
without use of brute force mechanics which we want to avoid if possible.

We need to finish writing the basic implementation of insert.

Co-authored-by: JAVIER MARIN <marinj5@rpi.edu>