Improve ArithmeticView implementation and functionalities
Added conversion function for infix expressions and implemented bindings for calculator's buttons. The ability to evaluate and display the history of expressions and corresponding evaluation results has also been incorporated. Changes to GUI include the addition of text fields and modification of welcome messages, along with comprehensive code enhancements to improve overall functionality of the application.