Updated UI styling and added new features
Modified the CSS classes for UI elements to include border styling in the index.js file. Added a new SVG container in the response section to hold a fox image. In the index.html, the Help text box has been changed to be contained in an accordion for ease of access. Also, added the fox SVG to the memory file system in the DefaultView.cpp file.