Add support for Integral function
This commit introduces support for the Integral function in the serialization and parsing of mathematical expressions. Fixes have been made to the `InFixSerializer` and `MathMLSerializer` classes to accommodate the new function. Furthermore, the `FromString` class has been updated to recognize and correctly parse the 'in' token as the Integral function.