Update GUI with additional mathematical symbols and layout options
This commit adds additional mathematical symbols, specifically integration and imaginary symbols, to the keypad in the DefaultView. It also updates the layout of the keypad to optimize space with the use of a wxGridBagSizer class. Furthermore, it adds new layout options in the "View" menu to allow changing between Horizontal and Vertical layouts. The new layout option is decided at runtime, enhancing the user’s customization capability.