[Feature] BoundedExpression, BoundedBinaryExpression, and BoundedUnaryExpression (#105)
* Add TernaryExpression and DefiniteIntegral to Oasis project

The update includes the addition of new modules, namely TernaryExpression and DefiniteIntegral, to the Oasis project. This is done by adding TernaryExpression.hpp, DefiniteIntegral.hpp and DefiniteIntegral.cpp files. Moreover, references to these files are also added in the CMakeLists.txt file to ensure proper building.

* Relocated IsAnyOf concept to Expression.hpp

The IsAnyOf concept, initially in BinaryExpression.hpp, has been moved to Expression.hpp. This relocation improves the organization of the code, seeing as the concept may have wider applicability than initially thought.

* [Actions] Run clang-format

* Trigger workflow

* Removed redefinition of IsAnyOf

* Refactor code to use Bounded expressions

This commit introduces BoundedUnaryExpression and BoundedBinaryExpression as supersets of TernaryExpression. It also replaces DefiniteIntegration with BoundedExpression, and adds a new type of expression, Summation. The changes include modification of associated header and CMakeLists files.

* [Actions] Run clang-format

* [Actions] Format CMakeLists.txt

* Trigger workflow

* Refactor Summation to use BoundedBinaryExpression

Summation class is now aliased to BoundedBinaryExpression instead of BoundedUnaryExpression, and includes an additional parameter for the incrementing variable. Additionally, updated the included headers in BoundedBinaryExpression.

---------

Co-authored-by: github-actions <github-actions@github.com>