Update Valgrind installation process in workflow
The way Valgrind is installed for profiling in the GitHub workflow is modified. Instead of using apt-get to install from the Ubuntu repository, Valgrind is now fetched directly from the source and compiled. This is controlled by a new flag matrix.build_profiling which only compiles Valgrind when the flag is set.