[Workflow] Make Valgrind actually work (#116)
* Enabled Debug and GCC builds

* Enabled g++

* Add Valgrind workflow

* Fix workflow

* Update method of artifact download in Valgrind workflow

The previous method of downloading artifacts in the GitHub Action for Valgrind testing has been updated. Instead of using 'actions/download-artifact@v4', it now uses 'actions/github-script@v5' to provide more direct and efficient artifact retrieval. It also includes an unzip operation for the artifact using built-in 'run' command.

* Remove code-coverage workflow, integrate coverage into multi-platform workflow

GitHub's code-coverage workflow has been removed to streamline the process. Instead of having a separate workflow for code coverage, it's now integrated into the multi-platform workflow. Coverage data and reports are only generated when the 'build_profiling' matrix option is set to true. Valgrind has also been added for memory profiling.

* Add Valgrind and memory check to GitHub workflow

The updates made ensure that the Valgrind package is installed during the workflow when in the 'ubuntu-latest' environment. Moreover, a memory check step has been added to the workflow under the 'build_profiling' matrix, using Valgrind for leak detection in Debug build config during the execution of 'OasisTests'.

* Remove valgrind.yml workflow file

This commit removes the valgrind.yml file from the GitHub workflows directory. We no longer need this workflow, as its functionality has been superseded and integrated into the general "CMake on multiple platforms" workflow. The code-coverage-comment.yml file has also been updated to trigger on this workflow instead of the previous "Code Coverage" workflow.

* Add reusable strings and PR number to GitHub workflow

The existing GitHub workflow file has been updated to include additional reusable strings for the coverage and PR directories. Furthermore, the PR number is now saved to facilitate subsequent commenting on the PR within the workflow process.

* Change command format in cmake-multi-platform.yml

The format of the command in the cmake-multi-platform.yml file has been changed. This was done to enhance readability and simplify the process of executing the command in different scenarios.

* Update Valgrind command in Github workflow

The Valgrind command in the cmake-multi-platform Github workflow has been updated. The `--build-config Debug` option was removed as it was unnecessary for the Valgrind memory leak check during the OasisTests run.

* Enable concurrent workflow runs in GitHub Actions

The updated settings in `cmake-multi-platform.yml` now allow for concurrent workflow runs in our GitHub Actions. This is controlled by the `concurrency` key which also cancels any in-progress runs if a new one starts. These changes aim to improve the efficiency of our CI process.

* Update Valgrind installation process in workflow

The way Valgrind is installed for profiling in the GitHub workflow is modified. Instead of using apt-get to install from the Ubuntu repository, Valgrind is now fetched directly from the source and compiled. This is controlled by a new flag matrix.build_profiling which only compiles Valgrind when the flag is set.

* Update valgrind install process in Github Actions

The tar extraction verbosity flag is modified to "-xvf" from "-xjf" for clear visibility during the workflow run. Also, the downloaded tar.bz2 file is removed post extraction to clean up disk space.

* Update cmake workflow to install libc6-dbg

This update modifies the cmake-multi-platform workflow under GitHub actions. In particular, it adds a command to update apt-get and install the libc6-dbg package when building with profiling enabled.

* Enabled Valgrind testing for GCC builds