Added a document detailing how to input mathematical expressions in Oasis
Basically I felt as though there should be at least some sort of documentation for how to use the project after you built it. If this sort of thing is not appropriate or should be more specific (ex. this is how you use the I/O for oasis, etc.) feedback for how to iterate would be appreciated.